Refactored tests, deleted duplicated ones
