+++ src/snowcli/cli/connection.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 from __future__ import annotations
 
 import typer
